Backend/requirements.txt
flask
ultralytics==8.3.33
torch==2.3.1
torchvision==0.18.1
pillow
numpy<2
